# More performance and code-size tweaks