Small code size improvement:

```
Title                              Section             Old             New  Percent
WebURLBenchmark                     __text:        1425601         1421009    -0.3%
```

Also some small performance improvements. The numbers are so tight that it's almost margin-of-error, but long paths going from 92ms to 78ms looks significant. We're hitting the point of diminishing returns, where even a 10% improvement amounts to ~2.5ms, which is difficult to reliably measure. In any case, this looks like it's moving in the right direction, the code looks nicer, and looking at the generated assembly, the gains seem realistic.

Before:

```
name                                                   time         std        iterations warmup         
---------------------------------------------------------------------------------------------------------
Constructor.SpecialNonFile.AverageURLs                 27093.000 ns ±  21.15 %      50830  6862462.000 ns
Constructor.SpecialNonFile.AverageURLs filtered        43105.000 ns ±  24.54 %      29515  4397805.000 ns
Constructor.SpecialNonFile.IPv4 host                   26586.000 ns ±  24.66 %      48909  2682248.000 ns
Constructor.SpecialNonFile.IPv4 host filtered          35014.000 ns ±  23.59 %      36598  3813488.000 ns
Constructor.SpecialNonFile.IPv6 host                   27508.000 ns ±  25.92 %      46141  3953910.000 ns
Constructor.SpecialNonFile.IPv6 host filtered          34646.500 ns ±  25.42 %      36586  3496124.000 ns
Constructor.SpecialNonFile.Percent-encoding components 11103.000 ns ±  28.59 %     120104  1128002.000 ns
Constructor.SpecialNonFile.Percent-encoded hostnames    9559.000 ns ±  25.98 %     137210  1029788.000 ns
Constructor.SpecialNonFile.Long paths                  92709.500 ns ±  21.19 %      13776 11634195.000 ns
Constructor.SpecialNonFile.Complex paths 1             26439.000 ns ±  23.22 %      49379  3268694.000 ns
Constructor.SpecialNonFile.Complex paths 2             32399.000 ns ±  24.43 %      39485  4699926.000 ns
Constructor.SpecialNonFile.Long query 1                 2445.000 ns ±  39.59 %     533375   321357.000 ns
Constructor.SpecialNonFile.Long query 2                21525.500 ns ±  23.29 %      58662  2854033.000 ns
```

After:

```
name                                                   time         std        iterations warmup        
--------------------------------------------------------------------------------------------------------
Constructor.SpecialNonFile.AverageURLs                 24860.000 ns ±  13.54 %      51832 3118901.000 ns
Constructor.SpecialNonFile.AverageURLs filtered        40453.000 ns ±  13.20 %      32265 4314659.000 ns
Constructor.SpecialNonFile.IPv4 host                   23958.000 ns ±  15.75 %      54709 2461829.000 ns
Constructor.SpecialNonFile.IPv4 host filtered          34855.000 ns ±  11.76 %      38040 3624207.000 ns
Constructor.SpecialNonFile.IPv6 host                   25037.000 ns ±  16.08 %      53182 2600285.000 ns
Constructor.SpecialNonFile.IPv6 host filtered          32421.000 ns ±  13.48 %      40545 3427260.000 ns
Constructor.SpecialNonFile.Percent-encoding components 10134.000 ns ±  16.55 %     119120 1148119.000 ns
Constructor.SpecialNonFile.Percent-encoded hostnames    8949.000 ns ±  14.63 %     149105  989243.000 ns
Constructor.SpecialNonFile.Long paths                  78507.000 ns ±   9.45 %      17017 8044446.000 ns
Constructor.SpecialNonFile.Complex paths 1             24138.000 ns ±   9.95 %      54900 2501131.000 ns
Constructor.SpecialNonFile.Complex paths 2             29436.000 ns ±  13.00 %      43328 3051703.000 ns
Constructor.SpecialNonFile.Long query 1                 2305.000 ns ±  27.23 %     545664  286796.000 ns
Constructor.SpecialNonFile.Long query 2                20301.000 ns ±  11.60 %      63377 2173905.000 ns
```